ml/requirements.txt
pymysql==1.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
jupyter==1.0.0
python-dotenv==1.0.0
joblib==1.3.2